+++ app/models/site.rb
@@ -4,7 +4,7 @@
 # members, and reports that are stratified by site.
 class Site < ApplicationRecord
   # Concerns
-  include Searchable, Deletable
+  include Searchable, Deletable, ShortNameable
 
   # Scopes
   def self.with_project_or_as_site_user(project_ids, user_id)
@@ -45,17 +45,6 @@ class Site < ApplicationRecord
 
   def subject_regex
     Regexp.new("\\A#{regex_string}\\Z") if regex_string.present?
-  end
-
-  def short_name
-    return self[:short_name] if self[:short_name].present?
-    computed_short_name
-  end
-
-  def computed_short_name
-    return name if name.to_s.split(/\s/).count <= 1
-    s = name.gsub(/(\b\w)([\w']*)/) { Regexp.last_match[1] }
-    s.to_s.gsub(/\s/, '')
   end
 
   def destroy