+++ db/migrate/20160923145736_add_short_name_to_treatment_arms.rb
@@ -1,0 +1,5 @@
+class AddShortNameToTreatmentArms < ActiveRecord::Migration[5.0]
+  def change
+    add_column :treatment_arms, :short_name, :string
+  end
+end